+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Eagerly call networker listen, remove await from swarm.join call #193
 
 ## [0.6.3] - 2020-07-14
 ### Changed
 - rename dat:// to hyper:// #169
-- pretty print JSON write-out #183 
+- pretty print JSON write-out #183
 - SDK now emits EBUSY error as a warn event #185
 
 ### Fixed